+++ packages/apollo/src/hooks/useMutation.ts
@@ -1,3 +1,3 @@
-import { useMutation } from '@apollo/react-hooks';
+import { useMutation } from 'react-apollo';
 
 export default useMutation;