+++ packages/apollo/src/hooks/useSubscription.ts
@@ -1,3 +1,3 @@
-import { useSubscription } from '@apollo/react-hooks';
+import { useSubscription } from 'react-apollo';
 
 export default useSubscription;