+++ packages/apollo/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.3.0 - 2019-08-01
+
+#### 🚀 Updates
+
+- Upgrade `react-apollo` to latest beta. Support native hooks. (#143) ([3cc66f7](https://github.com/airbnb/lunar/tree/master/packages/apollo/commit/3cc66f7)), closes [#143](https://github.com/airbnb/lunar/tree/master/packages/apollo/issues/143)
+
+**Note:** Version bump only for package @airbnb/lunar-apollo
+
+
+
+
+
 ### 2.2.1 - 2019-07-24
 
 **Note:** Version bump only for package @airbnb/lunar-apollo